video-sink: Ensure uploaded frame are displayed
This is fixed by making the uploading source equen priority to
the redraws. This is important, as Texture2D upload on slow memory
system can take a lot of time. You still want your UI to be
responsive (being able to show controls, resize, etc.) and not spend all
the time uploading. Ideally, we want exactly one upload per render but
we don't have that much control. A lower priority makes the video jitter
when there is overlayed animation happening.

https://bugzilla.gnome.org/show_bug.cgi?id=759208